Extended workercpu test program with reconos os call benchmarks. Here are the results, comparing the original fsl_driver and reconos lib with optimized versions.
			mbox put mbox get rqueue send	rqueue receive	semaphore	mutex	fsl single write fsl single read fsl block write fsl block read	memory write	memory read
original		1106.19	 1103.32  51.01		53.78		994.04		681.83	57.2		 60.22		 1885.78	 2025.46   	4146.8		9996.34
fsl_optimized		1145.1	 1093.49  50.87		51.77		980.67		680.07	56.5		 58.24		 3038.19	 2734.38   	4140.51		9978.08
fsl+reconos optimized	1100.54	 1140.53  617.12	593.5		982.87		682.8	56.55		 57.75		 3038.19	 2734.38   	4143.65		9996.34

This has been measured with Linux on a MicroBlaze as Master CPU and a Microblaze with bare-metal software as hardware thread (hwt). The MBOX and RQUEUE sizes are set to 1. RQUEUE message length is 1024 bytes.
FSL single writes and reads single words in a loop. FSL block write and read operate on 32KB of data. As a comparison, memory write and read show how fast the Microblaze Master Processor can read from /write
to main memory.
The units are as follows: mbox and rqueue numbers are calls per second on the master cpu. Semaphore and Mutex numbers are lock/unlock pairs per second. This is need for proper synchronization with slave.
The remaining numbers are kilobytes per second.

Normalized to the original version, improvements may be seen more easily:

			mbox put mbox get rqueue send	rqueue receive	semaphore	mutex	fsl single write fsl single read fsl block write fsl block read	memory write	memory read
original		1	 1	  1		1		1		1	1		 1		 1		 1	   	1		1
fsl_optimized		1.035	 0.991	  0.997		0.962		0.986		0.997	0.987		 0.967		 1.611		 1.350	   	0.998		0.998
fsl+reconos optimized	0.994	 1.033	  12.098	11.035		0.988		1.001	0.988		 0.958		 1.611		 1.350	   	0.999		1

In summary: Block transfers over fsl improved by 35% for reading and 61% for writing.
This resulted in a factor 11 to 12 boost for the reconos queue. All other calls have
nearly the same performance as before. Please note that i have only made one benchmark
run per version, so there might be some uncertainity.