ReconOS version 3.1
IMPORTANT: Please read this commit message before going on.

This commit includes the new ReconOS version 3.1 with a partially new
architecture and improved features, for example

  - moved from the PLB to the AXI bus
  - support for the Xilinx Zynq platform including very unique features
    like cache coherency, very stable Linux operating system, improved
    performance for software applications and reduced synthesis times
  - architecture independet library
  - support for dynamically reconfigurable hardware threads
  - a simple to use toolchain
  - support for more than 14 hardware threads (currently tested up to 31
    but in principal unlimited)

This commit removes all v3.0 stuff from the repository including the old
libraries and demo applications. Therefore, the state of the repository
before this commit can be found under the tag v3.0 and development with
version 3.0 is still possible in the appropriate branch. To upgrade your
currently running application, take a look into the migration guide
under docs/upgrade_30_31.txt. You will see, that only minimal changes
are needed to use applications running with ReconOS v3.0 on v3.1. Feel
free to migrate old demo applications or even develop new ones.

As far as it is tested, version 3.1 includes nearly all features of
version 3.0, except the follwing:

  - the rqueue is currently not implemented
  - the kernel space implementation of ReconOS is not available

To get an impression of the new ReconOS version, you can carry out the
step-by-step-guide in the docs directory. It includes detailed
instructions to build the sort-demo application for the Zynq (Zedboard)
or the microblaze (ML605).