Added script for analysis of Xilinx essential bit files. Essential bit files represent the important bits of a specific design, i.e. changing these bits leads to errors with a high probability.
The script translates the essential bits into physical addresses which can be used
for fault injection. Additionally it creates an image and tries to map the
essential bits to the physical FPGA layout. However, it is known how address elements half,
row and column map to the FPGA, while minor, word and bit are speculative.